GPT_WA_INTEGRATION/requirements.txt
anyio==3.7.1
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
distro==1.8.0
dnspython==2.2.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==2.2.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.26.2
openai==1.3.6
openpyxl==3.0.10
pandas==2.1.4
pandas-stubs==1.5.2.221213
pycodestyle==2.10.0
pydantic==1.10.13
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tqdm==4.64.1
twilio==7.16.0
types-pytz==2022.6.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
